PL2Q Hugin: Arm SCAD file now has option for 1 or 2 peice.  New geometries.
Printing the arm vertically takes a long time if your Z layer changing rate is slow.  Printing the tube in a horizontal fashion can also be tricky because it is a cylinder with lots of overhangs.  You can now choose - for aesthetics or practicality - the number of faces for your arm: 6, 7, 8, 11 or 30.